Custome Exception, DuplicateCustomerRequasteException were added, it will accure when trying to post from postman the same customer requeste

+++ bank/src/main/java/com/example/bank/customer/entity/CustomerEntity.java
@@ -23,7 +23,7 @@ public class CustomerEntity {
     @Temporal(TemporalType.DATE)
     private Date birthDate;
     @Column(name = "age",nullable = false)
-    private Byte age;
+    private Integer age;
     @Column(name = "phone", nullable = false, length = 32)
     private String phone;
     @Column(name = "email", nullable = false, length = 32)
@@ -96,11 +96,11 @@ public class CustomerEntity {
         this.birthDate = birthDate;
     }
 
-    public Byte getAge() {
+    public Integer getAge() {
         return age;
     }
 
-    public void setAge(final Byte age) {
+    public void setAge(final Integer age) {
         this.age = age;
     }
